+++ Level0-Module5/src/_02_nested_loops/_2_for_loop_gauntlet/Recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/index.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/ForLoopGauntlet.html"}
 </script>
