+++ Level0-Module5/src/_03_method_writing/_1_obedient_robot/Recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/index.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/ObedientRobot.html"}
 </script>
