+++ Level0-Module5/src/_04_practice/_1_skill_practice/SkillPractice.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/index.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/SkillPractice.html"}
 </script>
