+++ Level0-Module5/src/_05_self_tests/_1_which_data_type_1/Recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/index.html"}
+window.onload=function(){document.location.href="http://level0.jointheleague.org/levels/Level0/Module5/dataTpes1.html"}
 </script>
