+++ Level0-Module5/src/_05_self_tests/_5_for_loop_onslaught/Recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/index.html"}
+window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/forLoop)nslaught/index.html"}
 </script>
