+++ Level0-Module5/src/_05_self_tests/_6_level0_summary/Recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/index.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Module5/StudentSummary.html"}
 </script>
