+++ Level0-Module5/src/_05_self_tests/_8_online_practice_test1/Recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/index.html"}
+window.onload=function(){document.location.href="http://level0.jointheleague.org/Module5/onlinePractice.html"}
 </script>
